+++ lib/plausible/stats/imported/imported.ex
@@ -18,8 +18,8 @@ defmodule Plausible.Stats.Imported do
   Usually, when no filters are used, the imported schema supports the
   query. There is one exception though - breakdown by a custom property.
   We are currently importing only two custom properties - `url` and `path`.
-  Both these properties can only be used with their special goal filter
-  (see Plausible.Goals.SystemGoals).
+  Both these properties can only be used with their particular `event:name`
+  filter or the corresponding goal filter (see Plausible.Event.SystemEvents).
   """
   def schema_supports_query?(query) do
     length(Imported.Base.decide_tables(query)) > 0